Add function for button that will display number
I had to put the constant for selecting all buttons under the function
that creates the buttons in the first place.

Right now the display updates with one number at a time. Need to find a
way to add multiple and have that amount get stored as a value to be
used later.