Calculate two numbers
The function to calculate two numbers works, but that is all that works.
Any instances of more than one operator cause the numbers to join in
unexpected ways. 5+5+5 gives me a result of 555.

I need to find a way to split the whole string again or begin to look at
different ways of initially updating the calculator's display.