Variables update display & other buttons wired up
I was using the string of the entire display ('4+4') and then
attempting to split it up by numbers and operator. Now the display is
updated with the variables (num1 op num2).

Any operator button clicked after the initial variable setup will
calculate the result, turn that result into the first number, turn the
operator clicked into the current operator, and await input of the
second number.

The clear button will refresh the page. The = button will perform the
calculate function.